# docs/api-documentation.yaml
openapi: 3.0.0
info:
  title: MCP Advisor API
  description: API documentation for MCP Advisor - A discovery & recommendation service for Model Context Protocol servers
  version: 1.0.0
  contact:
    name: Yoshio Liu
    url: https://github.com/istarwyh/mcpadvisor
  license:
    name: MIT
    url: https://github.com/istarwyh/mcpadvisor/blob/main/LICENSE

servers:
  - url: http://localhost:3000
    description: Default local server
  - url: https://api.example.com
    description: Production server (example)

tags:
  - name: Health
    description: Health check endpoints
  - name: SSE
    description: Server-Sent Events endpoints
  - name: REST
    description: REST API endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Check if the server is running
      tags:
        - Health
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: 1.0.0
  
  /sse:
    get:
      summary: SSE connection
      description: Establish a Server-Sent Events connection
      tags:
        - SSE
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /messages:
    post:
      summary: Send message to SSE connection
      description: Send a message to an established SSE connection
      tags:
        - SSE
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
          description: The session ID of the SSE connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcRequest'
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /rest:
    post:
      summary: Direct REST API request
      description: Make a direct request to the MCP Advisor API
      tags:
        - REST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcRequest'
      responses:
        '200':
          description: Request processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    JsonRpcRequest:
      type: object
      required:
        - jsonrpc
        - method
        - params
        - id
      properties:
        jsonrpc:
          type: string
          enum: ['2.0']
          description: JSON-RPC version
        method:
          type: string
          enum: ['listTools', 'callTool']
          description: Method to call
        params:
          type: object
          oneOf:
            - $ref: '#/components/schemas/ListToolsParams'
            - $ref: '#/components/schemas/CallToolParams'
        id:
          type: string
          description: Request ID
      example:
        jsonrpc: '2.0'
        method: 'callTool'
        params:
          name: 'recommend-mcp-servers'
          arguments:
            query: 'financial data analysis'
        id: '1'
    
    ListToolsParams:
      type: object
      description: Parameters for listTools method
      
    CallToolParams:
      type: object
      required:
        - name
        - arguments
      properties:
        name:
          type: string
          description: Tool name
          enum: ['recommend-mcp-servers']
        arguments:
          type: object
          required:
            - query
          properties:
            query:
              type: string
              description: Search query for MCP servers
    
    JsonRpcResponse:
      type: object
      required:
        - jsonrpc
        - result
        - id
      properties:
        jsonrpc:
          type: string
          enum: ['2.0']
          description: JSON-RPC version
        result:
          oneOf:
            - $ref: '#/components/schemas/ListToolsResult'
            - $ref: '#/components/schemas/CallToolResult'
        id:
          type: string
          description: Request ID
      example:
        jsonrpc: '2.0'
        result:
          content:
            - title: 'Financial Analytics MCP'
              description: 'Comprehensive financial data analysis toolkit'
              github_url: 'https://github.com/example/financial-mcp'
              similarity: 0.95
        id: '1'
    
    ListToolsResult:
      type: object
      required:
        - tools
      properties:
        tools:
          type: array
          items:
            $ref: '#/components/schemas/Tool'
    
    Tool:
      type: object
      required:
        - name
        - description
        - inputSchema
      properties:
        name:
          type: string
          description: Tool name
        description:
          type: string
          description: Tool description
        inputSchema:
          type: object
          description: JSON Schema for tool input
    
    CallToolResult:
      type: object
      required:
        - content
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MCPServer'
    
    MCPServer:
      type: object
      required:
        - title
        - description
        - github_url
        - similarity
      properties:
        title:
          type: string
          description: Server title
        description:
          type: string
          description: Server description
        github_url:
          type: string
          format: uri
          description: GitHub URL for the server
        similarity:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Similarity score (0-1)
    
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: integer
              description: Error code
            message:
              type: string
              description: Error message
      example:
        error:
          code: 404
          message: 'Session not found'